refactor(operator_commands_dashboard): extract traces/activity + workboard handlers
Move traces and activity HTTP handlers into a new activity.rs sibling, and
move the workboard handler into a new workboard.rs sibling.

Promotes is_pid_alive and read_tail to pub(crate) for cross-module use.

- routes.rs: 4827 -> 4395
- activity.rs (NEW, 160)
- workboard.rs (NEW, 290)

Refs #1266
